+++ CHANGELOG.md
@@ -1,6 +1,22 @@
 ## Upcoming
 
 * #80 - Introspection descriptions for scalars and introspection
+* #78 - `Projection` and `NoProjection` should be tags instead of resolve function wrappers. Backwards-incompatible change: you need to replace
+  `Projection` with `ProjectionName` tag and `NoProjection` with `ProjectionExclude` tag. here is an example:
+  ```scala
+  // before
+
+  Field("id", StringType,
+    Some("The id of the droid."),
+    resolve = Projection("_id", _.value.id)),
+
+  // after
+
+  Field("id", StringType,
+    Some("The id of the droid."),
+    tags = ProjectionName("_id") :: Nil,
+    resolve = _.value.id),
+  ```
 
 ## v0.4.0 (2015-09-27)